- changed XText grammar: 	1. a Constant is now either a DecimalConstant or IntegerConstant 	2. align and length attribute are now of integer instead of BigDecimal 	3. Parameter no longer allow for attributes after value, i.e., allowing "i32 signext 42" but no longer "i32 42 signext" 	4. Return instruction may return either Parameter or PrimitiveValue 	5. GlobalDefinition.value can be either Parameter or Predefined
- refactored our Labeling and Graph utilities and actually removed lots
of duplicate code
(Seriously, who the f... needs 5 or more methods converting values to a
string representation????)

- TODO: transformations are broken right now and need adaption to the
changes

+++ de.upb.llvm_parser/src-gen/de/upb/llvm_parser/llvm/Alloc.java
@@ -2,7 +2,6 @@
  */
 package de.upb.llvm_parser.llvm;
 
-import java.math.BigDecimal;
 
 /**
  * <!-- begin-user-doc -->
@@ -112,12 +111,12 @@ public interface Alloc extends Instruction
    * </p>
    * <!-- end-user-doc -->
    * @return the value of the '<em>Align</em>' attribute.
-   * @see #setAlign(BigDecimal)
+   * @see #setAlign(int)
    * @see de.upb.llvm_parser.llvm.LlvmPackage#getAlloc_Align()
    * @model
    * @generated
    */
-  BigDecimal getAlign();
+  int getAlign();
 
   /**
    * Sets the value of the '{@link de.upb.llvm_parser.llvm.Alloc#getAlign <em>Align</em>}' attribute.
@@ -127,6 +126,6 @@ public interface Alloc extends Instruction
    * @see #getAlign()
    * @generated
    */
-  void setAlign(BigDecimal value);
+  void setAlign(int value);
 
 } // Alloc
