+++ de.upb.llvm_parser/src-gen/de/upb/llvm_parser/llvm/ExtractValue.java
@@ -2,8 +2,6 @@
  */
 package de.upb.llvm_parser.llvm;
 
-import java.math.BigDecimal;
-
 import org.eclipse.emf.common.util.EList;
 
 /**
@@ -80,7 +78,7 @@ public interface ExtractValue extends Instruction
 
   /**
    * Returns the value of the '<em><b>Index</b></em>' attribute list.
-   * The list contents are of type {@link java.math.BigDecimal}.
+   * The list contents are of type {@link java.lang.Integer}.
    * <!-- begin-user-doc -->
    * <p>
    * If the meaning of the '<em>Index</em>' attribute list isn't clear,
@@ -92,6 +90,6 @@ public interface ExtractValue extends Instruction
    * @model unique="false"
    * @generated
    */
-  EList<BigDecimal> getIndex();
+  EList<Integer> getIndex();
 
 } // ExtractValue
